First iteration on moving to shadcn style components
