+++ packages/headless-components/media/src/core/index.tsx
@@ -1,1 +1,2 @@
+export * as FileUpload from "./FileUpload.js";
 export * as MediaGallery from "./MediaGallery.js";
