+++ packages/headless-components/media/src/react/WixMediaImage.tsx
@@ -1,0 +1,68 @@
+import { Image, type FittingType, initCustomElement } from '@wix/image';
+
+type MediaItem = {
+  image?: string;
+};
+
+initCustomElement();
+
+const parseMediaFromUrl = (url: string) => {
+  const wixImagePrefix = 'wix:image://v1/';
+
+  if (url.startsWith(wixImagePrefix)) {
+    const uri = url.replace(wixImagePrefix, '').split('#')![0]!.split('/')[0]!;
+
+    const params = new URLSearchParams(url.split('#')[1] || '');
+    const originalWidth = params.get('originWidth');
+    const originalHeight = params.get('originHeight');
+
+    return {
+      uri,
+      originalWidth: originalWidth ? parseInt(originalWidth, 10) : undefined,
+      originalHeight: originalHeight ? parseInt(originalHeight, 10) : undefined,
+    };
+  }
+
+  throw new Error('Invalid image URL');
+};
+
+export function WixMediaImage({
+  media,
+  width,
+  height,
+  className,
+  alt = '',
+  showPlaceholder = true,
+  displayMode = 'fill',
+}: {
+  media?: MediaItem;
+  width?: number;
+  height?: number;
+  className?: string;
+  alt?: string;
+  showPlaceholder?: boolean;
+  displayMode?: FittingType;
+}) {
+  const {
+    uri,
+    originalWidth = 640,
+    originalHeight = 320,
+  } = parseMediaFromUrl(media?.image!);
+
+  return (
+    <Image
+      key={uri}
+      uri={uri}
+      width={width || originalWidth}
+      height={height || originalHeight}
+      containerWidth={width}
+      containerHeight={height}
+      displayMode={displayMode}
+      isSEOBot={false}
+      shouldUseLQIP={showPlaceholder}
+      alt={alt}
+      className={className}
+    />
+  );
+}
+
